% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/verbs.R
\name{select_.tbl_chunk}
\alias{filter}
\alias{filter_.tbl_chunk}
\alias{select_.tbl_chunk}
\title{Supported dplyr verbs}
\usage{
\method{select_}{tbl_chunk}(.data, ..., .dots)

\method{filter_}{tbl_chunk}(.data, ..., .dots)
}
\description{
\code{chunked} implements to following dplyr verbs
}

