% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/insert_chunk_into.R
\name{insert_chunkwise_into}
\alias{insert_chunkwise_into}
\title{insert data in chunks into a database}
\usage{
insert_chunkwise_into(x, dest, table, temporary = FALSE, analyze = FALSE)
}
\arguments{
\item{x}{tbl_chunk object}

\item{dest}{database destination, e.g. src_sqlite()}

\item{table}{name of table}

\item{temporary}{Should the table be removed when the database connection is closed?}

\item{analyze}{Should the table be analyzed after import?}
}
\value{
a \code{\link{tbl}} object pointing to the table in database \code{dest}.
}
\description{
\code{insert_chunkwise_into} can be used to insert chunks of data into a
database. Typically \code{chunked} can be used to for preprocessing data
before adding it to a database.
}

