% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_csv_chunkwise}
\alias{read_csv2_chunkwise}
\alias{read_csv_chunkwise}
\alias{read_table_chunkwise}
\title{Read data from text files chunkwise}
\usage{
read_csv_chunkwise(file, chunk_size = 5000L, header = TRUE, sep = ",",
  dec = ".", ...)

read_csv2_chunkwise(file, chunk_size = 5000L, header = TRUE, sep = ";",
  dec = ",", ...)

read_table_chunkwise(file, chunk_size = 5000L, header = TRUE, sep = "\\t",
  dec = ".", ...)
}
\arguments{
\item{file}{path of texst file}

\item{chunk_size}{size of the chunks te be read}

\item{header}{Does the csv file have a header with column names?}

\item{sep}{field separator to be used}

\item{dec}{decimal separator to be used}

\item{...}{not used}
}
\description{
read_csv_chunk will open a connection to a text file. Subsequent dplyr verbs and commands
are recorded until \code{collect} or \code{\link{write_csv_chunkwise}} is called. In that case
the recorded commands will be executed chunk by chunk.
}
\details{
read_csv_chunkwise can be best combined with write_csv_chunkwise (see example)
}

