% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.R
\name{write_chunkwise}
\alias{write_chunkwise}
\title{Genereric function to write chunk by chunk}
\usage{
write_chunkwise(x, dest, ...)
}
\arguments{
\item{x}{chunked input, e.g. created with \code{read_chunkwise}}

\item{dest}{where should the data be written. May be a character or
a \code{src_sql}.}

\item{...}{parameters that will be passed to the specific implementations.}
}
\description{
Genereric function to write chunk by chunk
}

