% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pkg.R
\docType{package}
\name{chunked-package}
\alias{chunked-package}
\alias{dplyr-verbs}
\alias{filter}
\alias{mutate}
\alias{select}
\title{Chunked}
\description{
With chunked you can use dplyr chunkwise on text files
}
\section{Implemented dplyr verbs}{


filter, select, do, left_join, inner_join
}

\section{Not implemented}{


 full_join, right_join, group_by, arrange, tail
}

