% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R
\name{read_csv_chunks}
\alias{read_csv2_chunks}
\alias{read_csv_chunks}
\alias{read_table_chunks}
\title{Read data from text files chunkwise}
\usage{
read_csv_chunks(file, chunk_size = 5000L, header = TRUE, sep = ",",
  dec = ".", ...)

read_csv2_chunks(file, chunk_size = 5000L, header = TRUE, sep = ";",
  dec = ",", ...)

read_table_chunks(file, chunk_size = 5000L, header = TRUE, sep = "\\t",
  dec = ".", ...)
}
\arguments{
\item{file}{path of texst file}

\item{chunk_size}{size of the chunks te be read}
}
\description{
read_csv_chunk will open a connection to a text file. Subsequent dplyr verbs and commands
are recorded until \code{collect} or \code{write_csv_chunks} is called. In that case
the recorded commands will be executed chunk by chunk.
}
\details{
read_csv_chunks can be best combined with write_csv_chunks (see example)
}

