% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.R
\name{write_chunkwise}
\alias{write_chunkwise}
\alias{write_chunkwise.chunkwise}
\title{Genereric function to write chunk by chunk}
\usage{
write_chunkwise(x, dest, ...)

\method{write_chunkwise}{chunkwise}(x, dest, table, file = dest,
  format = c("csv", "csv2", "table"), ...)
}
\arguments{
\item{x}{chunked input, e.g. created with \code{read_chunkwise} or it can be a \code{tbl_sql} object.}

\item{dest}{where should the data be written. May be a character or
a \code{src_sql}.}

\item{...}{parameters that will be passed to the specific implementations.}

\item{table}{table to write to. Only used when dest is a data base(\code{src_sql})}

\item{file}{File to write to}

\item{format}{Specifies the text format for written to disk. Only used
if \code{x} is a character.}
}
\description{
Genereric function to write chunk by chunk
}
