% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/write.R
\name{write_csv_chunkwise}
\alias{write_csv2_chunkwise}
\alias{write_csv_chunkwise}
\title{Write chunks to a csv file}
\usage{
write_csv_chunkwise(x, file = "", sep = ",", dec = ".",
  col.names = TRUE, row.names = FALSE, ...)

write_csv2_chunkwise(x, file = "", sep = ",", dec = ".",
  col.names = TRUE, row.names = FALSE, ...)
}
\arguments{
\item{x}{tbl_chunk object pointing to a text file}

\item{file}{file \code{character} or connection where the csv file should be written}

\item{sep}{field separator}

\item{dec}{decimal separator}

\item{col.names}{should column names be written?}

\item{row.names}{should row names be written?}

\item{...}{passed through to \code{\link{read.table}}}
}
\description{
Writes data to a csv file chunk by chunk.
}

